+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/qc/DetailedQcItem.java
@@ -1,5 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.core.data.qc;
 
+import java.util.Date;
+
 import com.eaglegenomics.simlims.core.User;
 
 import uk.ac.bbsrc.tgac.miso.core.data.DetailedQcStatus;
@@ -17,5 +19,9 @@ public interface DetailedQcItem {
   public User getQcUser();
 
   public void setQcUser(User qcUser);
+
+  public Date getQcDate();
+
+  public void setQcDate(Date qcDate);
 
 }