+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultRunPartitionAliquotService.java
@@ -72,7 +72,8 @@ public class DefaultRunPartitionAliquotService implements RunPartitionAliquotSer
       runPartitionAliquot.setPurpose(runPurposeService.get(runPartitionAliquot.getPurpose().getId()));
     }
     User user = authorizationManager.getCurrentUser();
-    updateQcUser(runPartitionAliquot, managed);
+    ValidationUtils.updateQcDetails(runPartitionAliquot, managed, RunPartitionAliquot::getQcPassed, RunPartitionAliquot::getQcUser,
+        RunPartitionAliquot::setQcUser, authorizationManager, RunPartitionAliquot::getQcDate, RunPartitionAliquot::setQcDate);
 
     List<ValidationError> errors = new ArrayList<>();
     ValidationUtils.validateQcUser(runPartitionAliquot.getQcPassed(), runPartitionAliquot.getQcUser(), errors);
@@ -91,20 +92,9 @@ public class DefaultRunPartitionAliquotService implements RunPartitionAliquotSer
       managed.setQcPassed(runPartitionAliquot.getQcPassed());
       managed.setQcNote(runPartitionAliquot.getQcNote());
       managed.setQcUser(runPartitionAliquot.getQcUser());
+      managed.setQcDate(runPartitionAliquot.getQcDate());
       managed.setLastModifier(user);
       runPartitionAliquotDao.update(managed);
-    }
-  }
-
-  private void updateQcUser(RunPartitionAliquot runPartitionAliquot, RunPartitionAliquot beforeChange) throws IOException {
-    if (ValidationUtils.isChanged(RunPartitionAliquot::getQcPassed, runPartitionAliquot, beforeChange)) {
-      if (runPartitionAliquot.getQcPassed() == null) {
-        runPartitionAliquot.setQcUser(null);
-      } else {
-        runPartitionAliquot.setQcUser(authorizationManager.getCurrentUser());
-      }
-    } else if (beforeChange != null) {
-      runPartitionAliquot.setQcUser(beforeChange.getQcUser());
     }
   }