+++ pinery-miso/src/it/resources/db/migration/pinery-miso_plain_test_data.sql
@@ -4,22 +4,22 @@ INSERT INTO Project(projectId, name, alias, shortName, created, description,
   (2, 'PRO2', 'Project Two', 'PRO2', '2017-06-27', 'integration test project two', 'ACTIVE', 1, '2017-06-27 14:11:00', 1, 1, 2);
 
 INSERT INTO Sample (sampleId, name, alias, description, identificationBarcode, sampleType, project_projectId,
-scientificNameId, volume, detailedQcStatusId, lastModifier, creator, created, lastModified, discriminator) VALUES
-(1, 'SAM1', 'TEST_0001', 'Identity', '11111', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2016-07-20 09:00:00', '2016-07-20 09:00:00', 'Sample'),
-(2, 'SAM2', 'TEST_0001_Bn_R_nn_1-1', 'Tissue', '22222', 'GENOMIC', 1, 1, 30, 1, 1, 1, '2016-07-20 09:01:00', '2016-07-20 09:01:00', 'Sample'),
-(3, 'SAM3', 'TEST_0001_Bn_R_nn_1-1_SL01', 'Slide', '33333', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(4, 'SAM4', 'TEST_0001_Bn_R_nn_1-1_C01', 'Curls', '44444', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(5, 'SAM5', 'TEST_0001_Bn_R_nn_1-1_LCM01', 'LCM Tube', '55555', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(6, 'SAM6', 'TEST_0001_Bn_R_nn_1-1_D_S1', 'gDNA stock', '66666', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(7, 'SAM7', 'TEST_0001_Bn_R_nn_1-1_R_S1', 'whole RNA stock', '77777', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(8, 'SAM8', 'TEST_0001_Bn_R_nn_1-1_D_1', 'gDNA aliquot', '88888', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(9, 'SAM9', 'TEST_0001_Bn_R_nn_1-1_R_1', 'whole RNA aliquot', '99999', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(10, 'SAM10', 'TEST_0001_Bn_R_nn_1-1_D_S2', 'cDNA stock', '10101', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(11, 'SAM11', 'TEST_0001_Bn_R_nn_1-1_D_2', 'cDNA aliquot', '11011', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(12, 'SAM12', 'TEST_0001_Bn_R_nn_1-1_R_1_SM_1', 'smRNA', '12121', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(13, 'SAM13', 'TEST_0001_Bn_R_nn_1-1_R_1_MR_1', 'mRNA', '13131', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(14, 'SAM14', 'TEST_0001_Bn_R_nn_1-1_R_1_WT_1', 'rRNA_depleted', '14141', 'GENOMIC', 1, 1, NULL, 1, 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
-(15, 'SAM15', 'PRO2_0001', 'Identity', '15151', 'GENOMIC', 2, 1, NULL, 1, 1, 1, '2017-07-20 09:00:00', '2017-07-20 09:00:00', 'Sample');
+scientificNameId, volume, detailedQcStatusId, qcDate, lastModifier, creator, created, lastModified, discriminator) VALUES
+(1, 'SAM1', 'TEST_0001', 'Identity', '11111', 'GENOMIC', 1, 1, NULL, 1, '2016-07-20', 1, 1, '2016-07-20 09:00:00', '2016-07-20 09:00:00', 'Sample'),
+(2, 'SAM2', 'TEST_0001_Bn_R_nn_1-1', 'Tissue', '22222', 'GENOMIC', 1, 1, 30, 1, '2016-07-20', 1, 1, '2016-07-20 09:01:00', '2016-07-20 09:01:00', 'Sample'),
+(3, 'SAM3', 'TEST_0001_Bn_R_nn_1-1_SL01', 'Slide', '33333', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(4, 'SAM4', 'TEST_0001_Bn_R_nn_1-1_C01', 'Curls', '44444', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(5, 'SAM5', 'TEST_0001_Bn_R_nn_1-1_LCM01', 'LCM Tube', '55555', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(6, 'SAM6', 'TEST_0001_Bn_R_nn_1-1_D_S1', 'gDNA stock', '66666', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(7, 'SAM7', 'TEST_0001_Bn_R_nn_1-1_R_S1', 'whole RNA stock', '77777', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(8, 'SAM8', 'TEST_0001_Bn_R_nn_1-1_D_1', 'gDNA aliquot', '88888', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(9, 'SAM9', 'TEST_0001_Bn_R_nn_1-1_R_1', 'whole RNA aliquot', '99999', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(10, 'SAM10', 'TEST_0001_Bn_R_nn_1-1_D_S2', 'cDNA stock', '10101', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(11, 'SAM11', 'TEST_0001_Bn_R_nn_1-1_D_2', 'cDNA aliquot', '11011', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(12, 'SAM12', 'TEST_0001_Bn_R_nn_1-1_R_1_SM_1', 'smRNA', '12121', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(13, 'SAM13', 'TEST_0001_Bn_R_nn_1-1_R_1_MR_1', 'mRNA', '13131', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(14, 'SAM14', 'TEST_0001_Bn_R_nn_1-1_R_1_WT_1', 'rRNA_depleted', '14141', 'GENOMIC', 1, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:01:00', '2017-07-20 09:01:00', 'Sample'),
+(15, 'SAM15', 'PRO2_0001', 'Identity', '15151', 'GENOMIC', 2, 1, NULL, 1, '2017-07-20', 1, 1, '2017-07-20 09:00:00', '2017-07-20 09:00:00', 'Sample');
 
 INSERT INTO Transfer(transferId, transferTime, senderLabId, recipientGroupId, creator, created, lastModifier, lastModified) VALUES
 (1, '2017-07-20 12:00:00', 1, 1, 1, '2017-07-20 12:53:00', 1, '2017-07-20 12:53:00');
@@ -34,10 +34,10 @@ INSERT INTO SampleChangeLog(sampleId, columnsChanged, message, userId, changeTim
 (2,'two','change twotwo',1,'2016-07-20 09:00:01');
 
 INSERT INTO Library(libraryId, name, alias, identificationBarcode, description, sample_sampleId, platformType,
-  libraryType, librarySelectionType, libraryStrategyType, creationDate, creator, created, lastModifier, lastModified, detailedQcStatusId, dnaSize,
+  libraryType, librarySelectionType, libraryStrategyType, creationDate, creator, created, lastModifier, lastModified, detailedQcStatusId, qcDate, dnaSize,
   volume, concentration, locationBarcode, kitDescriptorId, discriminator) VALUES
   (1, 'LIB1', 'TEST_0001_Bn_R_PE_300_WG', '11211', 'description lib 1', 8, 'ILLUMINA', 1, 3, 1,  '2016-11-07', 1, '2017-07-20 09:01:00',
-    1, '2017-07-20 09:01:00', 1, 300, 5.0, 2.75, NULL, 1, 'Library');
+    1, '2017-07-20 09:01:00', 1, '2017-07-20', 300, 5.0, 2.75, NULL, 1, 'Library');
 
 INSERT INTO Library_Index(library_libraryId, index_indexId) VALUES
   (1, 5);