+++ pinery-miso/src/main/resources/ca/on/oicr/pinery/lims/miso/queryAllRunSamples.sql
@@ -7,6 +7,7 @@ SELECT DISTINCT Run.runId
   ,COALESCE(Run.dataManglingPolicy, im.dataManglingPolicy) dataManglingPolicy
   ,rp.alias run_purpose
   ,rpa.qcPassed qc_passed
+  ,rpa.qcDate qc_date
 FROM SequencingContainerModel scm
 JOIN SequencerPartitionContainer spc ON scm.sequencingContainerModelId = spc.sequencingContainerModelId
 JOIN _Partition part ON part.containerId = spc.containerId